Use `nonReactiveCopy` instead of _.clone in CardEditor.vue
For example, we needed a code like this to copy `this.internalValue.fields`
instead of mutating it directly, because `this.internalValue.fields` is
reactive. However, because `_.cloneDeep` ** won't ** create a non-reactive
clone, we need a special `nonReactiveCopy` function.

```
const newFields = _.clone(this.internalValue.fields);
resize(newFields, fieldFormats.length, '');
this.internalValue.fields = newFields;
```